% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiercv.R
\name{ArbitraryDeptCF}
\alias{ArbitraryDeptCF}
\title{Critical Function for Hierarchical FDR Controlling Procedure under Arbitrary Dependence.}
\usage{
ArbitraryDeptCF(isTested, mi, li, l, depth, gdi, alpha, rOffset)
}
\arguments{
\item{isTested}{logical; if  \code{TRUE} (default), then the i-th hypotheses \eqn{H_i} will be tested; otherwise, \eqn{H_i} will not be tested.}

\item{mi}{the cardinality of the set of descendant hypotheses of \eqn{H_i}.}

\item{li}{the number of leaf hypotheses in the set of descendant hypotheses \eqn{H_i}.}

\item{l}{the total number of leaf hypotheses.}

\item{depth}{the cardinality of the set of ancestor hypotheses of \eqn{H_i}, referred to as the depth of \eqn{H_i}.}

\item{gdi}{the cardinality of the union set of all hypotheses with all depth no deeper than the given depth.}

\item{alpha}{the significant level used to calculate the critical values to make decisions.}

\item{rOffset}{the offset increment for the number of rejections.}
}
\value{
A critical function of the index \eqn{i} and rejection number \eqn{R}.
}
\description{
Given some parameters of the descendant hypotheses and ancestor hypotheses, return the critical function using the generalized hierarchical FDR controlling procedure under arbitrary dependence (see Theorem 2 in Lynch et al. (2016)).
}
\references{
Lynch, G., Guo, W. (2016).
 On Procedures Controlling the FDR for Testing Hierarchically Ordered Hypotheses.
 \emph{arXiv preprint} arXiv:1612.04467.
}
\author{
Yalin Zhu
}
