% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiercv.R
\name{hierFDR.CF}
\alias{hierFDR.CF}
\title{Critical Function for all Hierarchical FDR Controlling Procedure under Various Types of Dependence.}
\usage{
hierFDR.CF(isTested, mi, li, l, depth, fdi, gdi, alpha, rOffset, type)
}
\arguments{
\item{isTested}{logical; if  \code{TRUE} (default), then the i-th hypotheses \eqn{H_i} will be tested; otherwise, \eqn{H_i} will not be tested.}

\item{mi}{the cardinality of the set of descendant hypotheses of \eqn{H_i}.}

\item{li}{the number of leaf hypotheses in the set of descendant hypotheses \eqn{H_i}.}

\item{l}{the total number of leaf hypotheses.}

\item{depth}{the cardinality of the set of ancestor hypotheses of \eqn{H_i}, referred to as the depth of \eqn{H_i}.}

\item{fdi}{the cardinality of the set of all hypotheses with the given depth.}

\item{gdi}{the cardinality of the union set of all hypotheses with all depth no deeper than the given depth.}

\item{alpha}{the significant level used to calculate the critical values to make decisions.}

\item{rOffset}{the offset increment for the number of rejections.}

\item{type}{the type of dependence structure of the hierarchically ordered hypotheses. Currently, we provide four types of dependence: \code{"positive"}, \code{"arbitrary"}, \code{"block positive"} and \code{"block arbitrary"}.}
}
\value{
A critical function of the index \eqn{i} and rejection number \eqn{R}.
}
\description{
Given some parameters of the descendant hypotheses and ancestor hypotheses, return the critical function using the generalized hierarchical FDR controlling procedure under various type of dependence.
}
\references{
Lynch, G., Guo, W. (2016).
 On Procedures Controlling the FDR for Testing Hierarchically Ordered Hypotheses.
 \emph{arXiv preprint} arXiv:1612.04467.
}
\seealso{
\code{\link{PositiveDeptCF}}, \code{\link{ArbitraryDeptCF}}, \code{\link{BlockPositiveCF}}, \code{\link{BlockArbitraryCF}}
}
\author{
Yalin Zhu
}
