% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiercv.R
\name{stepwise}
\alias{stepwise}
\title{Decision-making Function for Generalized Stepwise Procedures}
\usage{
stepwise(p, CF, k)
}
\arguments{
\item{p}{numeric vector of p-values (possibly with \code{\link[base]{NA}}s). Any other R is coerced by \code{\link[base]{as.numeric}}. Same as in \code{\link[stats]{p.adjust}}.}

\item{CF}{critical functions of the corresponding multiple testing procedure}

\item{k}{the order of a generalized stepwise procedure.}
}
\value{
logical values of each hypothesis being rejected or not, if \code{TRUE}, then the hypothesis is rejected; otherwise, the hypothesis is not rejected.
}
\description{
Given a set of p-values, return the decisions using the generalized stepwise procedure.
}
\references{
Lynch, G., Guo, W. (2016).
 On Procedures Controlling the FDR for Testing Hierarchically Ordered Hypotheses.
 \emph{arXiv preprint} arXiv:1612.04467.

 Tamhane, A. C., Liu, W., & Dunnett, C. W. (1998).
 A generalized step-up-down multiple test procedure.
 \emph{The Canadian Journal of Statistics/La Revue Canadienne de Statistique}, 353-363.

 Sarkar, S. K. (2002).
 Some results on false discovery rate in stepwise multiple testing procedures.
 \emph{Annals of statistics}, 239-257.
}
\author{
Yalin Zhu
}
